+++ src/quantum_exercises/__init__.py
@@ -1,5 +1,5 @@
 """qx: a command line course in quantum computing with Qiskit."""
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 
 __all__ = ["__version__"]